// src/app/demo/models/modelsproyecto/actividadviewmodel.ts
export class Actividad{
    codigo?:number;
    acti_Id?: number;
    acti_Descripcion?: string;
    usua_Creacion?: number;
    usuaCreacion?: string;
    acti_FechaCreacion?: Date;
    fechaCreacion?: string;
    usua_Modificacion?: number;
    usuaModificacion?: string;
    acti_FechaModificacion?: Date;
    fechaModificacion?: string;
    acti_Estado?: boolean;
}
 